Remove `-[EXPExpect applyMathcer:to]`
This method seems to add unnecessary complication to the calling convention and appears to only have originally existed to allow for Async testing. The `to:(NSObject **)actual` parameter in `-[EXPExpect applyMatcher:to:]` appears to only have existed for the purpose of being a flag to say “we have a value now” - this is no longer required with the fixes from the previous commit.