Add explicit cleanUpBlock to EXPBlockDefinedMatcher.
Some matchers like `EXPMatchers+notify` are currently using `failureMessageForTo` and `failureMessageForNotTo`
to signify the end of the expectation and perform any clean up such as unregistering for notifications.
This prevents the matcher from being reused. By adding an explicit cleanUpBlock the matcher can safely do
whatever clean up it needs to do and know that it is not going to be used again from that point.