Create separate framework test targets
When you attempt to build the Expecta-iOS framework scheme via the following command with the iOS 7.1 Simulator installed:

```
xcodebuild build -scheme 'Expecta-iOS'
```

You'll receive the following error:

```
xcodebuild: error: Failed to build project Expecta with scheme Expecta-iOS.
	Reason: The run destination iPad 2 is not valid for Running the scheme 'Expecta-iOS'.
```

This is due to the fact that the `Expecta-iOS` framework scheme is currently configured to run tests using the `Expecta-iOSTests` target. However, this is an incorrect configuration, as the `Expecta-iOSTests` target does not actually depend on the framework target, but rather depends on the static library target. As such, running the Expecta-iOSTests target will not actually test the framework target. To fix this, I've:

- Renamed the old test targets as `libExpectaTests` and `libExpectaTests-iOS`
- Created new test targets `ExpectaTests` and `ExpectaTests-iOS` for the framework targets, to ensure that both integration options (framework or static library) can be properly tested